+++ recipes/configure-nginx-proxy.rb
@@ -7,6 +7,15 @@ install_package('nginx')
 
 install_nginx_logrotate_customizations
 
+worker_procs = get_nginx_worker_procs
+
+template %Q|/etc/nginx/nginx.conf| do
+  source 'nginx.conf.erb'
+  variables({
+    worker_procs: worker_procs
+  })
+end
+
 template %Q|/etc/nginx/sites-enabled/default| do
   source 'nginx-proxy.conf.erb'
   manage_symlink_source true